refactor: GridState를 별도 파일로 분리 #43
